- Fuse FunctionBuilder dedup + container-map into one O(V log V) sweep,
  eliminating a tautological second sweep and an O(V²) per-view rescan.
- Delete int_const_val (Option<u64>): read constants through the int_const_u128
  SSoT; the lift-time const-fold coercions now fold I80/I128 consts too.
- Widen the whole SP-offset/size/address chain (SpExpr, AddrClass, stack_offsets
  side-table, StackArgs, the pattern + Python function_arg_stack APIs) to i128
  and delete int_const_i64; adopt single_value_output in the SP load classifier.
- Plus the prior behavior-identical micro-simplifications (ir/opt/py).

Gate: cargo test --workspace exit 0 + clippy --all-targets clean + pytest 870.

+++ crates/strider-ir/src/node/kind.rs
@@ -19,7 +19,7 @@ pub enum FunctionArgSource {
     /// space of the stack (typically the architecture's RAM space).
     Stack {
         space: rsleigh::VnSpace,
-        offset: i64,
+        offset: i128,
     },
 }
